Move RcStack to collections folder.
