+++ crates/core/src/formatting/collections/mod.rs
@@ -1,5 +1,7 @@
 mod fast_cell_map;
 mod graph_node;
+mod rc_stack;
 
 pub use fast_cell_map::*;
 pub use graph_node::*;
+pub use rc_stack::*;
